update user schema models
